apply patch #89843 from Jan Wolter to fix BQM sharp tool settings conversion to XML CCBUGS: 318726

+++ libs/dimg/filters/sharp/unsharpmaskfilter.cpp
@@ -204,7 +204,7 @@ FilterAction UnsharpMaskFilter::filterAction()
 void UnsharpMaskFilter::readParameters(const FilterAction& action)
 {
     m_amount    = action.parameter("amount").toDouble();
-    m_radius    = action.parameter("radius").toInt();
+    m_radius    = action.parameter("radius").toDouble();
     m_threshold = action.parameter("threshold").toDouble();
 }
